Update Quarkus example
With Quarkus 0.17.0 and newer, it is possible to use AMQP with the reactive messaging API based on microprofile. This PR replaces the quarkus and microprofile client examples with one based on the Quarkus [amqp quickstart](https://github.com/quarkusio/quarkus-quickstarts/tree/master/amqp-quickstart), with modifications to inject EnMasse endpoint information.